autopilot-agent/pyproject.toml
[tool.poetry]
name = "autopilot"
version = "0.1.0"
description = "Self-driving agent using Groq API directly"
authors = ["aruyaohi <omariontechnologies@gmail.com>"]
packages = [{ include = "autopilot", from = "src" }]

[tool.poetry.dependencies]
python = ">=3.11,<3.12"
groq = "^0.5.0"  # Official Groq SDK
python-dotenv = "^1.1.0"
waymo-waymax = { git = "https://github.com/waymo-research/waymax.git", branch = "main" }

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
